feat(api): Build alert rule list api (SEN-834)
This also includes the serializer for alert rules. I wanted to include these serializers in the
incidents app, but due to the way our serializers are structured we're unable to do so due to
circular imports. Ideally we could split up our `api/` app into `api_framework/` and `api/` so that
it's more portable, but I'll leave that for now.